requirements.txt
Flask
flask_sqlalchemy
weasyprint
reportlab
matplotlib